requirements.txt
networkx
matplotlib.pyplot
os